Get started my journey.
Here’s a try to build a simple statistic tool. And this is my first
piece of code contributed to Github. I don’t know where the road will
lead to and it maybe given up in the future. Because I’m not major in
CS or some alike major. It’s only a interest and I’m learning on it.